[ADD] - CD based on tags (#7)
* [ADD] - CD based on tags

If a tag is pushed on the repository, the docker image is deployed with the same tag version. The Latest version is also deployed (obviously)

* [EDIT] - Jobs renamed

Renamed "manual-deployment" to "tag-deployment". Step "Build and push" precised with the image name

* [EDIT] - tag-deployment refined

The tag-deployment job is reached only if the deployment-latest job is successful