[ADD] - Hadolint (#8)
* [ADD] - CD based on tags

If a tag is pushed on the repository, the docker image is deployed with the same tag version. The Latest version is also deployed (obviously)

* [EDIT] - Jobs renamed

Renamed "manual-deployment" to "tag-deployment". Step "Build and push" precised with the image name

* [EDIT] - tag-deployment refined

The tag-deployment job is reached only if the deployment-latest job is successful

* [ADD] - Hadolint

\/!\\ From this commit, the deployment-latest job runs for all pushes on master (cf. PR merged).

CI completed with an Hadolint test for the docker image

* [EDIT] - Syntax for hadolint step fixed

* [EDIT] - Syntax for hadolint step fixed

* [EDIT] - Syntax for hadolint step fixed

* [EDIT] - Syntax for hadolint step fixed

* [EDIT] - Syntax for hadolint step fixed

Trying to unset the name for the comment in PR0

* [EDIT] - Syntax for hadolint step fixed

* [EDIT] - remove the comment on PR

It doesn't work, leave an empty comment. Only a bonus so ciao